+++ pkg/api/dtos/models.go
@@ -32,6 +32,9 @@ type DashboardMeta struct {
 	IsStarred  bool      `json:"isStarred,omitempty"`
 	IsHome     bool      `json:"isHome,omitempty"`
 	IsSnapshot bool      `json:"isSnapshot,omitempty"`
+	Type       string    `json:"type,omitempty"`
+	CanSave    bool      `json:"canSave"`
+	CanStar    bool      `json:"canStar"`
 	Slug       string    `json:"slug"`
 	Expires    time.Time `json:"expires"`
 	Created    time.Time `json:"created"`